Renaming a couple things per code review, adding newline to end of files
