+++ src/js/flags/glitches.ts
@@ -1,4 +1,4 @@
-import {FlagSection} from './model';
+import {FlagSection} from './flag';
 
 export const glitchFlags: FlagSection = {
   section: 'Glitches',
