+++ src/js/flags/hard-mode.ts
@@ -1,4 +1,4 @@
-import {FlagSection} from './model';
+import {FlagSection} from './flag';
 
 export const hardModeFlags: FlagSection = {
   section: 'Hard mode',
